+++ python/acl_anthology/people/person.py
@@ -14,8 +14,9 @@
 
 from __future__ import annotations
 
-from attrs import define, field, Factory
+from attrs import define, field
 from typing import Iterator, Optional, TYPE_CHECKING
+from ..utils.attrs import auto_validate_types
 from ..utils.ids import AnthologyIDTuple, build_id_from_tuple
 from . import Name
 
@@ -24,9 +25,12 @@ if TYPE_CHECKING:
     from ..collections import Paper, Volume
 
 
-@define
+@define(field_transformer=auto_validate_types)
 class Person:
     """A natural person.
+
+    Info:
+        All information about persons is currently derived from [name specifications][acl_anthology.people.name.NameSpecification] on volumes and papers, and not stored explicitly. This means that Person objects **cannot be used to make changes** to Anthology data; change the information on papers instead.
 
     Attributes:
         id: A unique ID for this person.
@@ -37,14 +41,14 @@ class Person:
         is_explicit: True if this person has names explicitly defined in `name_variants.yaml`.  Note this does _not_ necessarily mean an explicit ID was defined for the person there.
     """
 
-    id: str
+    id: str = field()
     parent: Anthology = field(repr=False, eq=False)
-    names: list[Name] = Factory(list)
+    names: list[Name] = field(factory=list)
     item_ids: list[AnthologyIDTuple] = field(
         factory=list, repr=lambda x: f"<list of {len(x)} AnthologyIDTuple objects>"
     )
     comment: Optional[str] = field(default=None)
-    is_explicit: Optional[bool] = field(default=False)
+    is_explicit: Optional[bool] = field(default=False)  # TODO: why can this be None?
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, Person):