+++ python/acl_anthology/utils/latex.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""Functions implementing the conversion to LaTeX/BibTeX formats."""
+
 from __future__ import annotations
 
 import re