+++ python/tests/collections/collection_test.py
@@ -17,8 +17,10 @@ from lxml import etree
 from pathlib import Path
 
 from acl_anthology import Anthology
-from acl_anthology.collections import Collection, CollectionIndex
+from acl_anthology.collections import Collection, CollectionIndex, VolumeType
+from acl_anthology.people import NameSpecification
 from acl_anthology.utils import xml
+from acl_anthology.text import MarkupText
 
 
 test_cases_xml_collections = (
@@ -39,7 +41,7 @@ def collection_index(anthology):
 
 
 def test_empty_collection(collection_index):
-    collection = Collection("empty", parent=collection_index, path=Path("."))
+    collection = Collection("2000.empty", parent=collection_index, path=Path("."))
     assert not collection.is_data_loaded
     collection.is_data_loaded = True
     assert isinstance(collection.root, Anthology)
@@ -100,3 +102,193 @@ def test_collection_roundtrip_save(collection_index, datadir, tmp_path, filename
     expected = etree.parse(infile)
     generated = etree.parse(outfile)
     xml.assert_equals(generated.getroot(), expected.getroot())
+
+
+def test_collection_create_volume_implicit(collection_index):
+    collection = collection_index.get("2022.acl")
+    volume = collection.create_volume(
+        "keynotes",
+        title=MarkupText.from_string("Keynotes from ACL 2022"),
+    )
+    assert volume.id in collection
+    assert volume.year == "2022"
+    assert volume.id == "keynotes"
+    assert volume.full_id == "2022.acl-keynotes"
+    assert volume.type == VolumeType.PROCEEDINGS
+
+
+def test_collection_create_volume_explicit(collection_index):
+    collection = collection_index.get("J89")
+    collection.id = "1989.cl"  # because we can't create volumes with old-style ID
+    volume = collection.create_volume(
+        id="99",
+        title=MarkupText.from_string("Special Issue"),
+        year="1989",
+        type="journal",
+        journal_issue="99",
+        venue_ids=["cl"],
+    )
+    assert volume.id in collection
+    assert volume.year == "1989"
+    assert volume.id == "99"
+    assert volume.full_id == "1989.cl-99"
+    assert volume.type == VolumeType.JOURNAL
+    assert volume.journal_issue == "99"
+    assert "cl" in volume.venue_ids
+
+
+@pytest.mark.parametrize("pre_load", (True, False))
+def test_collection_create_volume_should_update_person(anthology, pre_load):
+    if pre_load:
+        anthology.people.load()  # otherwise we test creation, not updating
+
+    collection = anthology.collections.get("2022.acl")
+    editors = [NameSpecification("Rada Mihalcea")]
+    volume = collection.create_volume(
+        "keynotes",
+        title=MarkupText.from_string("Keynotes from ACL 2022"),
+        editors=editors,
+    )
+    assert volume.editors == editors
+
+    # Volume should have been added to the person object
+    person = anthology.resolve(editors[0])
+    assert volume.full_id_tuple in person.item_ids
+
+
+@pytest.mark.parametrize("pre_load", (True, False))
+def test_collection_create_volume_should_update_personindex(anthology, pre_load):
+    if pre_load:
+        anthology.people.load()  # otherwise we test creation, not updating
+
+    collection = anthology.collections.get("2022.acl")
+    editors = [NameSpecification("Nonexistant, Guy Absolutely")]
+    volume = collection.create_volume(
+        "keynotes",
+        title=MarkupText.from_string("Keynotes from ACL 2022"),
+        editors=editors,
+    )
+    assert volume.editors == editors
+
+    # New editor should exist in the person index
+    person = anthology.resolve(editors[0])
+    assert volume.full_id_tuple in person.item_ids
+
+
+@pytest.mark.parametrize(
+    "pre_load, reset",
+    (
+        (True, True),
+        pytest.param(True, False, marks=pytest.mark.xfail(reason="not implemented")),
+        (False, False),
+        (False, True),
+    ),
+)
+def test_collection_create_volume_should_create_event(anthology, pre_load, reset):
+    if pre_load:
+        anthology.events.load()  # otherwise we test creation, not updating
+
+    collection = anthology.collections.create("2000.empty")
+    volume = collection.create_volume(
+        "1",
+        title=MarkupText.from_string("Empty volume"),
+        venue_ids=["acl"],
+    )
+
+    if reset:
+        anthology.reset_indices()
+
+    # New implicit event should exist in the event index
+    assert "acl-2000" in anthology.events
+    assert volume.full_id_tuple in anthology.events["acl-2000"].colocated_ids
+    assert volume.full_id_tuple in anthology.events.reverse
+    assert anthology.events.reverse[volume.full_id_tuple] == {"acl-2000"}
+
+
+@pytest.mark.parametrize(
+    "pre_load, reset",
+    (
+        (True, True),
+        pytest.param(True, False, marks=pytest.mark.xfail(reason="not implemented")),
+        (False, False),
+        (False, True),
+    ),
+)
+def test_collection_create_volume_should_update_event(anthology, pre_load, reset):
+    if pre_load:
+        anthology.events.load()  # otherwise we test creation, not updating
+
+    collection = anthology.collections.get("2022.acl")
+    collection.is_data_loaded = True
+    volume = collection.create_volume(
+        "keynotes",
+        title=MarkupText.from_string("Keynotes from ACL 2022"),
+        venue_ids=["acl"],
+    )
+
+    if reset:
+        anthology.reset_indices()
+
+    # New volume should be added to existing event
+    assert "acl-2022" in anthology.events
+    assert volume.full_id_tuple in anthology.events["acl-2022"].colocated_ids
+    assert volume.full_id_tuple in anthology.events.reverse
+    assert anthology.events.reverse[volume.full_id_tuple] == {"acl-2022"}
+
+
+@pytest.mark.parametrize(
+    "pre_load, reset",
+    (
+        (True, True),
+        pytest.param(True, False, marks=pytest.mark.xfail(reason="not implemented")),
+        (False, False),
+        (False, True),
+    ),
+)
+def test_collection_create_volume_should_update_venue(anthology, pre_load, reset):
+    if pre_load:
+        anthology.venues.load()  # otherwise we test creation, not updating
+
+    collection = anthology.collections.create("2000.empty")
+    volume = collection.create_volume(
+        "1",
+        title=MarkupText.from_string("Empty volume"),
+        venue_ids=["acl"],
+    )
+
+    if reset:
+        anthology.reset_indices()
+
+    # Nev volume should be added to existing venue
+    assert volume.full_id_tuple in anthology.venues["acl"].item_ids
+
+
+def test_collection_create_event(collection_index):
+    collection = collection_index.get("L06")
+
+    # For old-style ID collections, an ID must be explicitly given
+    with pytest.raises(ValueError):
+        _ = collection.create_event()
+
+    event = collection.create_event(id="lrec-2006")
+    assert event.id == "lrec-2006"
+
+    # Trying to create yet another event in the same collection should raise
+    with pytest.raises(ValueError):
+        _ = collection.create_event(id="lrecagain-2006")
+
+
+@pytest.mark.parametrize("pre_load", (True, False))
+def test_collection_create_event_should_update_eventindex(pre_load, anthology):
+    if pre_load:
+        anthology.events.load()  # otherwise we test creation, not updating
+
+    collection = anthology.collections.get("L06")
+    event = collection.create_event(id="lrec-2006")
+
+    if pre_load:
+        # Volume should automatically have been added
+        assert event.colocated_ids == [collection.get("1").full_id_tuple]
+    else:
+        # If event index wasn't loaded, it's not
+        assert event.colocated_ids == []