+++ python/tests/collections/volume_test.py
@@ -18,6 +18,7 @@ from pathlib import Path
 import pytest
 
 from acl_anthology.collections import Collection, Volume, VolumeType
+from acl_anthology.people import NameSpecification as NameSpec
 from acl_anthology.text import MarkupText
 from acl_anthology.utils.xml import indent
 
@@ -136,9 +137,9 @@ def test_volume_minimum_attribs():
 def test_volume_all_attribs():
     volume_title = MarkupText.from_string("Lorem ipsum")
     volume_shorttitle = MarkupText.from_string("L.I.")
-    parent = Collection("2023.acl-long", None, Path("."))
+    parent = Collection("2023.acl", None, Path("."))
     volume = Volume(
-        id="42",
+        id="long",
         parent=parent,
         type="proceedings",
         booktitle=volume_title,
@@ -254,7 +255,7 @@ def test_volume_with_multiple_venues(anthology):
     volume_title = MarkupText.from_string(
         "Joint conference of ACL and LREC (hypothetical)"
     )
-    parent = Collection("acl.2092", CollectionIndexStub(anthology), Path("."))
+    parent = Collection("2092.acl", CollectionIndexStub(anthology), Path("."))
     volume = Volume(
         "1",
         parent,
@@ -263,7 +264,7 @@ def test_volume_with_multiple_venues(anthology):
         venue_ids=["acl", "lrec"],
         year="2092",
     )
-    assert volume.full_id == "acl.2092-1"
+    assert volume.full_id == "2092.acl-1"
     assert volume.title == volume_title
     assert volume.venue_ids == ["acl", "lrec"]
     assert volume.venue_acronym == "ACL-LREC"
@@ -294,3 +295,161 @@ def test_volume_roundtrip_xml(xml):
     out = volume.to_xml()
     indent(out)
     assert etree.tostring(out, encoding="unicode") == xml
+
+
+def test_volume_generate_paper_id(anthology):
+    volume = anthology.get_volume("2022.acl-long")
+    # Highest paper ID in 2022.acl-long is 603
+    assert volume.generate_paper_id() == "604"
+    # Calling this repeatedly will generate the same ID
+    assert volume.generate_paper_id() == "604"
+    # Adding a Paper with this ID should then generate the next-higher one
+    volume.create_paper(
+        id="604",
+        bibkey="my-awesome-paper",
+        title=MarkupText.from_string("The awesome paper I have never written"),
+    )
+    assert volume.generate_paper_id() == "605"
+
+
+def test_volume_create_paper_implicit(anthology):
+    volume = anthology.get_volume("2022.acl-long")
+    authors = [NameSpec("Bollmann, Marcel")]
+    paper = volume.create_paper(
+        title=MarkupText.from_string("The awesome paper I have never written"),
+        authors=authors,
+        ingest_date="2025-01-07",
+    )
+    assert paper.authors == authors
+    assert paper.title.as_text() == "The awesome paper I have never written"
+    assert paper.ingest_date == "2025-01-07"
+    assert paper.parent is volume
+    assert paper.id in volume
+    # Highest paper ID in 2022.acl-long is 603, so this one should automatically get 604
+    assert paper.id == "604"
+    assert paper.full_id == "2022.acl-long.604"
+    # Bibkey should automatically have been generated
+    assert paper.bibkey == "bollmann-2022-awesome"
+
+
+def test_volume_create_paper_explicit(anthology):
+    volume = anthology.get_volume("2022.acl-long")
+    authors = [NameSpec("Bollmann, Marcel")]
+    paper = volume.create_paper(
+        title=MarkupText.from_string("The awesome paper I have never written"),
+        authors=authors,
+        ingest_date="2025-01-07",
+        id="701",
+        bibkey="bollmann-2022-the-awesome",
+    )
+    assert paper.authors == authors
+    assert paper.title.as_text() == "The awesome paper I have never written"
+    assert paper.ingest_date == "2025-01-07"
+    assert paper.parent is volume
+    assert paper.id in volume
+    assert paper.id == "701"
+    assert paper.full_id == "2022.acl-long.701"
+    assert paper.bibkey == "bollmann-2022-the-awesome"
+
+
+@pytest.mark.parametrize("pre_load", (True, False))
+def test_volume_create_paper_should_update_person(anthology, pre_load):
+    if pre_load:
+        anthology.people.load()  # otherwise we test creation, not updating
+
+    volume = anthology.get_volume("2022.acl-long")
+    authors = [NameSpec("Berg-Kirkpatrick, Taylor")]
+    paper = volume.create_paper(
+        title=MarkupText.from_string("The awesome paper I have never written"),
+        authors=authors,
+        ingest_date="2025-01-07",
+    )
+    assert paper.authors == authors
+
+    # Paper should have been added to the person object
+    person = anthology.resolve(authors[0])
+    assert paper.full_id_tuple in person.item_ids
+
+
+@pytest.mark.parametrize("pre_load", (True, False))
+def test_volume_create_paper_should_update_personindex(anthology, pre_load):
+    if pre_load:
+        anthology.people.load()  # otherwise we test creation, not updating
+
+    volume = anthology.get_volume("2022.acl-long")
+    authors = [NameSpec("Nonexistant, Guy Absolutely")]
+    paper = volume.create_paper(
+        title=MarkupText.from_string("An entirely imaginary paper"),
+        authors=authors,
+        ingest_date="2025-01-07",
+    )
+    assert paper.authors == authors
+
+    # New author should exist in the author index
+    person = anthology.resolve(authors[0])
+    assert paper.full_id_tuple in person.item_ids
+
+
+def test_volume_remove_editor(anthology):
+    volume = anthology.get_volume("2022.acl-long")
+    ns = volume.editors[1]
+    person = anthology.resolve(ns)
+    assert person.id == "preslav-nakov"
+    assert volume.full_id_tuple in person.item_ids
+
+    # Removing editor from volume
+    volume.editors = [volume.editors[0], volume.editors[2]]
+    # Person should be updated after resetting indices
+    anthology.reset_indices()
+    person = anthology.resolve(ns)
+    assert volume.full_id_tuple not in person.item_ids
+
+
+def test_volume_add_editor(anthology):
+    volume = anthology.get_volume("2022.acl-long")
+    # This person exists, but is not an editor on this volume
+    ns = NameSpec("Rada Mihalcea")
+    assert ns not in volume.editors
+    person = anthology.resolve(ns)
+    assert volume.full_id_tuple not in person.item_ids
+
+    # Adding this editor to the volume
+    volume.editors.append(ns)
+    # Person should be updated after resetting indices
+    anthology.reset_indices()
+    person = anthology.resolve(ns)
+    assert volume.full_id_tuple in person.item_ids
+
+
+def test_volume_type_conversion():
+    parent = Collection("L05", None, Path("."))
+    volume = Volume(
+        6,
+        parent,
+        type="journal",
+        booktitle=MarkupText.from_string("Lorem ipsum"),  # "Lorem ipsum",
+        year=2005,
+    )
+    assert volume.id == "6"  # str
+    assert volume.full_id == "L05-6"
+    assert isinstance(volume.title, MarkupText)
+    assert volume.year == "2005"
+    assert volume.type == VolumeType.JOURNAL
+
+
+def test_volume_type_validation():
+    volume_title = MarkupText.from_string("Lorem ipsum")
+    parent = Collection("L05", None, Path("."))
+    volume = Volume(
+        "6",
+        parent,
+        type=VolumeType.JOURNAL,
+        booktitle=volume_title,
+        year="2005",
+    )
+    with pytest.raises(TypeError):
+        volume.doi = 42
+    with pytest.raises(TypeError):
+        volume.venue_ids = "lrec"
+    with pytest.raises(TypeError):
+        volume.pdf = "L05-6000.pdf"