+++ python/CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Added
 
 - Support for Python 3.13.
+- Papers are now indexed by their bibkeys and can be retrieved via `Anthology.get_paper_by_bibkey()`.
+- Bibkeys can now be generated and updated, guaranteeing uniqueness.
+- Collections, Volumes, Papers, and Events can now be newly created with functions on their respective parent objects.
+  - Event creation currently has some unintuitive behaviour due to the existence of implicit event creation and linking; see docs.
+- FileReferences can now be instantiated from files, and functions for checksum computation have been added.
+
+### Changed
 
 ## [0.5.1] — 2025-01-02